pyproject.toml
[tool.poetry]
name = "dlplcr9000evm"
version = "0.1.0"
description = "Python package for controlling the DLP LightCrafter 9000 EVM (Evaluation Module)."
authors = ["Jakub Filip <mr.jakubfilip@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.24.3"
pillow = "^9.5.0"
hidapi = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
